refactor: move filter components to own component
